refactor(tiledmap): 重构Token。将BaseToken与ImageToken区分
